Separate file for MockDoseStore
